+++ rnd_mwc_l5.c
@@ -121,7 +121,7 @@ double rnd_double(struct rnd *rnd)
 double rnd_double_2(struct rnd *rnd)
 {
 /* Return double (0,2) in continuous triangular distribution */
-	return OPEN_DBL(next(rnd)+next(rnd));
+	return OPEN_DBL((double)next(rnd)+(double)next(rnd));
 }
 
 double rnd_double_n(struct rnd *rnd, unsigned n)
